refactor: extract the grid layout from wallets and contacts to \`WalletGrid\`
