+++ src/renderer/components/Wallet/index.js
@@ -1,4 +1,5 @@
 import WalletAddress from './WalletAddress'
+import WalletGrid from './WalletGrid'
 import WalletIdenticon from './WalletIdenticon'
 import WalletIdenticonPlaceholder from './WalletIdenticonPlaceholder'
 import WalletRemovalConfirmation from './WalletRemovalConfirmation'
@@ -21,6 +22,7 @@ export * from './WalletButtons'
 
 export {
   WalletAddress,
+  WalletGrid,
   WalletIdenticon,
   WalletIdenticonPlaceholder,
   WalletRemovalConfirmation,
